requirements.txt
moviepy<2.0.0
pillow
pylint
black
pytube
tqdm
PyQt6